Rename fritzdect package to fritz, API to Meter, SwitchAPI to Switch
- Rename meter/fritzdect -> meter/fritz
- Rename fritzdect.go -> legacy.go (legacy LUA API implementation)
- Rename API interface to Meter
- Rename SwitchAPI interface to Switch
- Update all imports and references
