+++ meter/fritz/legacy.go
@@ -1,4 +1,4 @@
-package fritzdect
+package fritz
 
 import (
 	"crypto/md5"
